requirements.txt
flask>=2.0.0
torch>=1.9.0
sentence-transformers>=2.2.0
google-generativeai
gunicorn==20.1.0